% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{checkIfNas}
\alias{checkIfNas}
\title{Chequea si existen nulos en el data.frame de entrada y da errores de validacion.}
\usage{
checkIfNas(X_1, X, vars)
}
\arguments{
\item{X_1}{data.frame convertido al tipo de datos correcto sin validar.}

\item{X}{data.frame con las variables independientes sin validar.}

\item{vars}{Lista con la descripcion de las variables que entran en el modelo.}
}
\description{
Chequea si existen nulos en el data.frame de entrada y da errores de validacion.
}
