% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{checkMissingColumns}
\alias{checkMissingColumns}
\title{Chequea si existen todas las columnas en el data.frame de entrada. Sino
lanza un error de validacion.}
\usage{
checkMissingColumns(X, vars)
}
\arguments{
\item{X}{data.frame con las variables independientes sin validar.}

\item{vars}{Lista con la descripcion de las variables que entran en el modelo.}
}
\description{
Chequea si existen todas las columnas en el data.frame de entrada. Sino
lanza un error de validacion.
}
