% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{coerceData}
\alias{coerceData}
\title{Convierte un data.frame a los tipos de datos soportados por el modelo. Si no
es posible lanza errores de validacion.}
\usage{
coerceData(X, vars)
}
\arguments{
\item{X}{data.frame con las variables independientes sin validar.}

\item{vars}{Lista con la descripcion de las variables que entran en el modelo.}
}
\value{
data.frame validado y convertido a los datos de entrada al modelo.
}
\description{
Convierte un data.frame a los tipos de datos soportados por el modelo. Si no
es posible lanza errores de validacion.
}
