% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{convertData}
\alias{convertData}
\title{Convierte el data.frame de entrada al tipo de datos del modelo.}
\usage{
convertData(X, vars)
}
\arguments{
\item{X}{data.frame con las variables independientes sin validar.}

\item{vars}{Lista con la descripcion de las variables que entran en el modelo.}
}
\value{
data.frame con las columnas en el tipo de datos correcto, aunque
pueden producirse registros nulos.
}
\description{
Convierte el data.frame de entrada al tipo de datos del modelo.
}
