% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{inputFeatures}
\alias{inputFeatures}
\title{Función que devuelve la descripción de las variables de entrada como una
lista. El nombre para cada objeto de la lista debe ser el nombre de la
variable y debe tener un atributo 'class' que puede ser 'numeric' o 'factor'.
En el caso de ser 'factor' tiene que contener un atributo extra 'levels',
que contiene los niveles de cada atributo.}
\usage{
inputFeatures(mdl)
}
\arguments{
\item{mdl}{Modelo de entrada.}
}
\value{
Lista de variables con las características de la descripción de la
función genérica.
}
\description{
Función que devuelve la descripción de las variables de entrada como una
lista. El nombre para cada objeto de la lista debe ser el nombre de la
variable y debe tener un atributo 'class' que puede ser 'numeric' o 'factor'.
En el caso de ser 'factor' tiene que contener un atributo extra 'levels',
que contiene los niveles de cada atributo.
}
