% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_generics.R
\name{inputFeatures.train}
\alias{inputFeatures.train}
\title{Implementación para un modelo de caret dónde se han guardado los datos de
entrenamiento del mismo.}
\usage{
\method{inputFeatures}{train}(mdl)
}
\arguments{
\item{mdl}{Modelo entrenado con caret. Tiene que contener el atributo
'trainingData'.}
}
\value{
Lista de variables con las características de la descripción de la
función genérica.
}
\description{
Implementación para un modelo de caret dónde se han guardado los datos de
entrenamiento del mismo.
}
