% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inputFeaturesFromDataFrame}
\alias{inputFeaturesFromDataFrame}
\title{Obtiene una descripci贸n de las variables en funci贸n de un data.frame en
particular.}
\usage{
inputFeaturesFromDataFrame(X)
}
\description{
Obtiene una descripci贸n de las variables en funci贸n de un data.frame en
particular.
}
