% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modelInfo}
\alias{modelInfo}
\title{Función que devuelve información sobre el modelo a modo de una lista de
strings.}
\usage{
modelInfo(mdl)
}
\arguments{
\item{mdl}{Objeto que contiene el modelo.}
}
\value{
Lista de strings con información sobre el modelo.
}
\description{
Función que devuelve información sobre el modelo a modo de una lista de
strings.
}
