% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_generics.R
\name{modelInfo.train}
\alias{modelInfo.train}
\title{Implementación para un modelo de caret. Opcionalmente se le puede añadir
los atributos 'name' y 'version' para llevar un control de los modelos.}
\usage{
\method{modelInfo}{train}(mdl)
}
\arguments{
\item{mdl}{Objeto que contiene el modelo.}
}
\value{
Lista de strings con información sobre el modelo.
}
\description{
Implementación para un modelo de caret. Opcionalmente se le puede añadir
los atributos 'name' y 'version' para llevar un control de los modelos.
}
