% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_model.R
\docType{data}
\name{PlumberMocker}
\alias{PlumberMocker}
\title{Añade la funcionalidad a un objeto plumber de poder simular requests desde
R sin tener que levantar la API. Ayuda para testeo.
Simula una petición a la api.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberMocker
}
\arguments{
\item{verb}{Verbo de peticion "GET", "POST" ...}

\item{path}{Url de la petición, por ejemplo '/foo'.}

\item{query_str}{Parametros como una url query 'foo=1&&bar=fooBar'.}

\item{body}{Cuerpo de la petición, por ejemplo un JSON.
Especializacion del metodo request para peticiones GET.
Especializacion del metodo request para peticiones POST.}
}
\description{
Añade la funcionalidad a un objeto plumber de poder simular requests desde
R sin tener que levantar la API. Ayuda para testeo.
Simula una petición a la api.
}
\keyword{datasets}
