% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_model.R
\docType{data}
\name{PlumberModel}
\alias{PlumberModel}
\title{Métodos públicos. Las clases hijas los deben implementar si el modelo es
distinto para ser compatibles con la api.
Constructor de la api.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberModel
}
\arguments{
\item{model}{Modelo base de la API.
Obtiene una copia del modelo base.}

\item{X}{data.frame con las variables independientes.}

\item{req}{Objeto request.}

\item{req}{Objeto de peticion.}

\item{res}{Objeto de respuesta.}

\item{err}{Objeto de error.}
}
\value{
Copia del modelo base.
Fija el modelo base.

Referencia a sí mismo.
Obtiene información sobre el modelo.

Lista con información básica sobre el modelo.
Resultados del entrenamiento.

data.frame con los resultados del entrenamiento en función de
los hiperparámetros del modelo.
Definición de las variables del modelo.

Lista con las variables e información sobre las mismas.
Predice el modelo.

Vector con las predicciones de la variable objetivo.
Define los endpoints de la API.
Obtiene los argumentos de la query url como un data.frame. Todas las
columnas son de caracteres.

data.frame con los argumentos.
Maneja los errores que se producen en los endpoints.
}
\description{
Métodos públicos. Las clases hijas los deben implementar si el modelo es
distinto para ser compatibles con la api.
Constructor de la api.
}
\keyword{datasets}
