% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plumberModelSkeleton}
\alias{plumberModelSkeleton}
\title{Crea el esqueleto de una aplicaci贸n.
Crea un bootstrap de la aplicaci贸n con un Dockerfile y scripts para correrlo
y crearlo.}
\usage{
plumberModelSkeleton(dir = ".")
}
\arguments{
\item{dir}{Directorio en el que crear el esqueleto.}
}
\description{
Crea el esqueleto de una aplicaci贸n.
Crea un bootstrap de la aplicaci贸n con un Dockerfile y scripts para correrlo
y crearlo.
}
