% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_model.R
\docType{data}
\name{PlumberModelWebApp}
\alias{PlumberModelWebApp}
\title{Añade a la API un servidor web de archivos estáticos que se monta sobre la
url '/'.
Constructor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberModelWebApp
}
\arguments{
\item{model}{Modelo sobre el que se construye la API.}

\item{static.dir}{Directorio desde el que se sirven los archivos.
estaticos.
Construye el servidor de archivos estáticos.}

\item{static.dir}{Directorio desde el que se sirven los archivos.}
}
\description{
Añade a la API un servidor web de archivos estáticos que se monta sobre la
url '/'.
Constructor
}
\keyword{datasets}
