% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{trainResults}
\alias{trainResults}
\title{Función que obtiene los resultados del entrenamiento de un modelo,
devolviendo una serie de métricas como un data.frame.}
\usage{
trainResults(mdl)
}
\arguments{
\item{mdl}{Objeto que contiene el modelo.}
}
\value{
data.frame con las métricas del modelo.
}
\description{
Función que obtiene los resultados del entrenamiento de un modelo,
devolviendo una serie de métricas como un data.frame.
}
