% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_generics.R
\name{trainResults.train}
\alias{trainResults.train}
\title{Implementación para un modelo de caret.}
\usage{
\method{trainResults}{train}(mdl)
}
\arguments{
\item{mdl}{Objeto que contiene el modelo.}
}
\value{
data.frame con columnas 'Metric' con el nombre de la metrica y
columna 'Value' con el valor de la métrica.
}
\description{
Implementación para un modelo de caret.
}
